# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dictionary API",
        "description": "A RESTful API for dictionary management with user authentication and word lookup",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Dictionary API Server"
        }
    ],
    "paths": {
...
        "/api/dictionary": {
            "get": {
                "tags": [
                    "Dictionary"
                ],
                "summary": "Search for words",
                "operationId": "9481954c1e1f5722b130cd9a3646a51d",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Word"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dictionary"
                ],
                "summary": "Create new dictionary entry",
                "operationId": "2f5cd02e9b745a0e925710606a530312",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "word": {
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "definition": {
                                        "type": "string",
                                        "example": "A sample definition"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "This is an example sentence."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dictionary entry created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Word"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dictionary/{word}": {
            "get": {
                "tags": [
                    "Dictionary"
                ],
                "summary": "Get word details",
                "operationId": "2ce0ac24ae8945387e887ec595d65cc8",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "description": "Word to look up",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Word"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Word not found"
                    }
                }
            }
        },
        "/api/dictionary/{word}/favorite": {
            "post": {
                "tags": [
                    "Dictionary"
                ],
                "summary": "Add word to favorites",
                "operationId": "82575acf9ab37e8163f29e32e92b8ed0",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "description": "Word to favorite",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Word added to favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Word added to favorites"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Word not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dictionary"
                ],
                "summary": "Remove word from favorites",
                "operationId": "283738d87f44ebcaddca327467e0636e",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "description": "Word to unfavorite",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Word removed from favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Word removed from favorites"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Word not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dictionary/{id}": {
            "get": {
                "tags": [
                    "Dictionary"
                ],
                "summary": "Get dictionary entry by ID",
                "operationId": "7ea721cc3532636c6b77480dee5df4b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dictionary entry details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Word"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dictionary entry not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Dictionary"
                ],
                "summary": "Update dictionary entry",
                "operationId": "d322ff7dea56f98b71a357eeb96991c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "word": {
                                        "type": "string",
                                        "example": "updated example"
                                    },
                                    "definition": {
                                        "type": "string",
                                        "example": "An updated definition"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "This is an updated example sentence."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dictionary entry updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Word"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dictionary entry not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dictionary"
                ],
                "summary": "Delete dictionary entry",
                "operationId": "ed542c213048cb15ac6703cd9c20f381",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Dictionary entry deleted successfully"
                    },
                    "404": {
                        "description": "Dictionary entry not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Word": {
                "required": [
                    "word",
                    "language"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "word": {
                        "type": "string",
                        "example": "fire"
                    },
                    "language": {
                        "type": "string",
                        "example": "en"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "totalDocs": {
                        "type": "integer",
                        "example": 100
                    },
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "totalPages": {
                        "type": "integer",
                        "example": 10
                    },
                    "hasNext": {
                        "type": "boolean",
                        "example": true
                    },
                    "hasPrev": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}